groupe-BESTFOREVER/backend/src/tools/activities.py

import os
from dotenv import load_dotenv

load_dotenv()

def find_activities(city: str) -> dict:
    """
    Finds activities for a given city.

    Args:
        city: The city to search for activities in.

    Returns:
        A dictionary containing the activity data, or an error message.
    """
    api_key = os.getenv("GOOGLE_MAPS_API_KEY")
    if not api_key:
        # Mocked data for now
        return {
            "activities": [
                {
                    "name": "Mock Museum",
                    "type": "museum"
                },
                {
                    "name": "Mock Park",
                    "type": "park"
                }
            ]
        }

    # TODO: Implement actual API call to Google Maps Places API or another similar API
    return {"error": "Activity search is not yet implemented with a real API."}

